% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOTB.R
\name{add2Path}
\alias{add2Path}
\title{Add a variable to the system path}
\usage{
add2Path(newPath)
}
\arguments{
\item{newPath}{Path to be added}
}
\value{
NONE
}
\description{
Add a variable to the system path
}
\examples{
\dontrun{
}
}
\author{
Chris Reudenbach
}

