% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRasterTiles.R
\name{combineRasterTiles}
\alias{combineRasterTiles}
\title{Tile \code{Raster*} object with or without overlap}
\usage{
combineRasterTiles(inpath, subpath, outpath, trgt)
}
\arguments{
\item{subpath}{'character'. Path created within each tile path (e.g. for naming the product).}

\item{outpath}{'character'. Top level path of the output files.}

\item{raster}{\code{Raster*}. Object that should be tiled.}

\item{tilenbr}{'integer'. Vector given the number of tiles in c(x,y) direction.}

\item{overlap}{'integer'. Overlap in pixels between the individual tiles. Default NULL does not include any overlap.}
}
\value{
None. Tiles are written to GeoTiff files in directory \code{outpath}.
}
\description{
Tile \code{Raster*} object using an optional overlap between the tiles.
}
\examples{
\\notrun{
combineRasterTiles(raster = ndvi, tilenbr = c(5,4), overlap = 30, outpath = modis_tiles)
}

}
\references{
NONE
}
