% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFromRasterSnips.R
\name{extractFromRasterSnips}
\alias{extractFromRasterSnips}
\title{Extract data from \code{Raster*}  snips based on selector from 
\code{Spatial*} object}
\usage{
extractFromRasterSnips(raster, spatial, selector = NULL, buffer = NULL)
}
\arguments{
\item{raster}{\code{Raster*} object}

\item{spatial}{\code{SpatialPointDataFrame} object}

\item{selector}{Selector variable, if not NULL, only one raster tile per 
unique value of the selector variable is created (e.g. if one has multiple
observations per single geographical location).}

\item{buffer}{Buffer size (default NULL does not use buffer) if the mean and
standard deviation within the buffer should be retruned, too.}
}
\value{
A \code{SpatialPointDataFrame} object with the values of the raster added to the attribute
table.
}
\description{
Extract data from \code{Raster*} snips based on selector from 
\code{Spatial*} objects. Aside from the actual raster values, the mean and
standard deviation of raster values within a buffer can be returned.

This function is usefull if the snips have been
computed by \code{\link{snipRaster}} using a selector ID since in this case,
the sequence of snips is not related to the sequence of \code{Spatial*}
entities in the vector layer.
}
\examples{
\\notrun{
}

}
\author{
Thomas Nauss
}
\references{

}

