% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOTB.R
\name{getOutputDir}
\alias{getOutputDir}
\title{Set outpath directory if not supplied for OTB functions}
\usage{
getOutputDir(path_output)
}
\description{
set output directory to working_directory/outpathOTB if not supplied for OTB functions
}
\examples{
\dontrun{
}
}

