% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTextures.R, R/sampleRasterFromPolygons.R
\docType{methods}
\name{imageTextures}
\alias{imageTextures}
\alias{imageTextures,RasterLayer-method}
\alias{imageTextures,RasterStack-method}
\alias{imageTextures,Satellite-method}
\alias{sampleRasterFromPolygons,RasterLayer-method}
\alias{sampleRasterFromPolygons,RasterStack-method}
\alias{sampleRasterFromPolygons,Satellite-method}
\title{Calculate image textures}
\usage{
\S4method{imageTextures}{Satellite}(x, bcde, mask = NULL, ...)

\S4method{imageTextures}{RasterStack}(x, mask = NULL, ...)

\S4method{imageTextures}{RasterLayer}(x, mask = NULL, ...)

\S4method{sampleRasterFromPolygons}{Satellite}(x, bcde, mask = NULL, ...)

\S4method{sampleRasterFromPolygons}{RasterStack}(x, poly, nbr = 50,
  res = raster::res(x), ...)

\S4method{sampleRasterFromPolygons}{RasterLayer}(x, poly, nbr = NULL, ...)
}
\arguments{
\item{x}{Satellite object, raster stack or raster layer.}

\item{bcde}{Band code(s) to be used for computing the image textures}

\item{mask}{raster layer with the same geometry as x with NA values at
positions where image textures should not be computed.}

\item{...}{Further arguments passed on to \code{\link{glcm::glcm}}}

\item{mask}{A \code{raster::RasterLayer} in which area of interest are marked
with 1, NA otherwise.}

\item{mask}{A \code{raster::RasterLayer} in which area of interest are marked
with 1, NA otherwise.}
}
\value{
If x is a Satellite object, a Satellite object with added image
textures; if x is a \code{raster::Raster*} object, a \code{raster::Raster*} 
object with converted layer(s).
}
\description{
Calculate image textures
}
\details{
The method is a wrapper for of glcm:glcm by A. Zvoleff
(see \url{https://cran.r-project.org/web/packages/glcm/index.html})
}
\examples{
\dontrun{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.TIF"), full.names = TRUE)
sat <- satellite(files)

sat <- imageTextures(sat, bcde = "B004n")
}
}
\references{
None.
}

