% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOTB.R
\name{initOTB}
\alias{initOTB}
\title{initOTB setup the orfeo toolbox bindings for an rsession}
\usage{
initOTB(defaultOTBPath = NULL, installationRoot = NULL, otbType = NULL,
  DL = "C:", otbselect = FALSE)
}
\arguments{
\item{defaultOTBPath}{string contains path to otb binaries}

\item{DL}{hard drive letter}
}
\value{
add otb pathes to the enviroment and creates global variables otbPath
}
\description{
initOTB trys to find all valid OTB installation and returns the pathes and environment settings
}
\details{
initOTB
}
\examples{
\dontrun{
## call it for a default OSGeo4W oinstallation of SAGA
initOTB("C:\\\\OSGeo4W64\\\\bin\\\\")
}
}
\author{
CR
}

