% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOTB.R
\name{makGlobalVar}
\alias{makGlobalVar}
\title{Assign variable in .GlobalEnv}
\usage{
makGlobalVar(name, value)
}
\arguments{
\item{name}{Variable name}

\item{value}{Value of varaible}
}
\value{
NONE
}
\description{
Assign variable in .GlobalEnv
}
\examples{
\dontrun{
}
}
\author{
Chris Reudenbach
}
