% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTiledRasters.R
\name{mergeTiledRasters}
\alias{mergeTiledRasters}
\title{Merge tiled \code{Raster*} objects back together}
\usage{
mergeTiledRasters(rasterlist, overlap = 0, outpath)
}
\arguments{
\item{rasterlist}{\code{Raster*}. List of raster objects that should be merged back together.}

\item{overlap}{'integer'. Overlap in pixels used for \code{tileRaster}}

\item{outpath}{'character'. Path and name of the output file.}
}
\value{
None. Merged tiles are written to single GeoTiff file \code{outpath}.
}
\description{
Merge tiled \code{Raster*} objects computed by \code{tileRaster}.
}
\examples{
\\notrun{
mergeTiledRasters(rasterlist = my_list, overlap = 30, outpath = merged_raster)
}

}
\references{
NONE
}
