% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpecIndices.R
\name{mSpecIndices}
\alias{mSpecIndices}
\title{Compute spectral remote sensing pixel-based indices}
\usage{
mSpecIndices(blue, green, red, nir)
}
\arguments{
\item{sat}{a satellite::satellite class data set}
}
\value{
a raster::stack of the indices
}
\description{
Compute spectral remote sensing pixel-based indices
}
\examples{
\dontrun{
}
}
