% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbcli_ExactLargScaleMeanShiftSegmentation.R
\docType{methods}
\name{otbcli_ExactLargeScaleMeanShiftSegmentation}
\alias{otbcli_ExactLargeScaleMeanShiftSegmentation}
\alias{otbcli_ExactLargeScaleMeanShiftSegmentation,character-method}
\title{Calculates a mean shift filtering (required for object based classification)}
\usage{
\S4method{otbcli_ExactLargeScaleMeanShiftSegmentation}{character}(x,
  inpos = NULL, out = NULL, tmpdir = NULL, spatialr = 5, ranger = 15,
  minsize = 0, tilesizex = 500, tilesizey = 500, verbose = FALSE,
  return_raster = FALSE)
}
\arguments{
\item{x}{A \code{Raster*} object or a gdall raster containing the filterd input file}

\item{inpos}{A \code{Raster*} object or a gdall raster containing the spatial input file}

\item{out}{The full path and name of the output file}

\item{tmpdir}{Path of temporary output files}

\item{spatialr}{Computational radius in euclidian distance (m)}

\item{ranger}{Range in multi-spectral space to be grouped into one}

\item{minsize}{Minimum size of the extracted segments}

\item{tilesizex}{Size of individual processing tiles in x direction}

\item{tilesizey}{Size of individual processing tiles in y direction}

\item{verbose}{switch for system messages default is FALSE}

\item{return_raster}{boolean if TRUE a raster stack is returned}
}
\description{
Calculates a mean shift filtering
}
\note{
the otb is used for the calculation of the statistics.
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbcli_ExactLargeScaleMeanShiftSegmentation(input=paste0(getwd(),"4490600_5321400.tif"),spatialr=5)
}
}
\author{
Thomas Nauss
}

