% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbcli_LSMSVectorization.R
\docType{methods}
\name{otbcli_LSMSVectorization}
\alias{otbcli_LSMSVectorization}
\alias{otbcli_LSMSVectorization,character-method}
\title{Calculates a mean shift filtering (required for object based classification)}
\usage{
\S4method{otbcli_LSMSVectorization}{character}(x, inseg = NULL, out = NULL,
  tilesizex = 500, tilesizey = 500, verbose = FALSE, ram = "8192")
}
\arguments{
\item{x}{A \code{Raster*} object or a gdall raster containing the original input file}

\item{inseg}{A \code{Raster*} object or a gdall raster containing the finally semgented  file}

\item{out}{The full path and name of the output file (shp)}

\item{tilesizex}{Size of individual processing tiles in x direction}

\item{tilesizey}{Size of individual processing tiles in y direction}

\item{verbose}{switch for system messages default is FALSE}

\item{ram}{reserved memory in MB}
}
\description{
Calculates a mean shift filtering
}
\note{
the otb is used for the calculation of the statistics.
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbcli_LSMSVectorization(input=paste0(getwd(),"4490600_5321400.tif"),spatialr=5)
}
}
\author{
Thomas Nauss
}

