% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbcli_MeanShiftSmoothing.R
\docType{methods}
\name{otbcli_MeanShiftSmoothing}
\alias{otbcli_MeanShiftSmoothing}
\alias{otbcli_MeanShiftSmoothing,RasterBrick-method}
\alias{otbcli_MeanShiftSmoothing,RasterLayer-method}
\alias{otbcli_MeanShiftSmoothing,RasterStack-method}
\alias{otbcli_MeanShiftSmoothing,character-method}
\title{Calculates a mean shift filtering (required for object based classification)}
\usage{
\S4method{otbcli_MeanShiftSmoothing}{RasterBrick}(x, path_temp = NULL,
  return_raster = TRUE, spatialr = 5, range = 15, thres = 0.1,
  maxiter = 100, rangeramp = 0, modesearch = "True", verbose = FALSE,
  ram = "8192")

\S4method{otbcli_MeanShiftSmoothing}{RasterLayer}(x, path_temp = NULL,
  return_raster = TRUE, spatialr = 5, range = 15, thres = 0.1,
  maxiter = 100, rangeramp = 0, modesearch = "True", verbose = FALSE,
  ram = "8192")

\S4method{otbcli_MeanShiftSmoothing}{RasterStack}(x, path_temp = NULL,
  return_raster = TRUE, spatialr = 5, range = 15, thres = 0.1,
  maxiter = 100, rangeramp = 0, modesearch = "True", verbose = FALSE,
  ram = "8192")

\S4method{otbcli_MeanShiftSmoothing}{character}(x, outfile_filter = NULL,
  outfile_spatial = NULL, return_raster = FALSE, spatialr = 5,
  range = 15, thres = 0.1, maxiter = 100, rangeramp = 0,
  modesearch = "True", verbose = FALSE, ram = "8192")
}
\arguments{
\item{x}{A \code{Raster*} object or a gdall raster containing 1 or more gray 
value bands}

\item{path_temp}{Path of temporary output files}

\item{return_raster}{boolean if TRUE a raster stack is returned}

\item{spatialr}{Computational radius in euclidian distance (m)}

\item{range}{Range in multi-spectral space to be grouped into one}

\item{thres}{Mode convergence threshold}

\item{maxiter}{Maximum number of iterations}

\item{rangeramp}{Spectral range coefficient}

\item{modesearch}{Mode search}

\item{verbose}{switch for system messages default is FALSE}

\item{ram}{reserved memory in MB}
}
\description{
Calculates a mean shift filtering
}
\note{
the otb is used for the calculation of the statistics.
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbcli_MeanShiftSmoothing(input=paste0(getwd(),"4490600_5321400.tif"),spatialr=5)
}
}
\author{
Thomas Nauss
}

