% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbHaraTex.R
\docType{methods}
\name{otbHaraTex}
\alias{otbHaraTex}
\alias{otbHaraTex,RasterBrick-method}
\alias{otbHaraTex,RasterLayer-method}
\alias{otbHaraTex,RasterStack-method}
\alias{otbHaraTex,character-method}
\title{Calculate selected texture parameters on gray level properties}
\usage{
\S4method{otbHaraTex}{RasterBrick}(x, path_output = NULL,
  return_raster = TRUE, ram = "8192", parameters.xyrad = list(c(1, 1)),
  parameters.xyoff = list(c(1, 1)), parameters.minmax = c(0, 255),
  parameters.nbbin = 8, texture = "advanced", channel = NULL,
  verbose = FALSE)

\S4method{otbHaraTex}{RasterLayer}(x, path_output = NULL,
  return_raster = TRUE, ram = "8192", parameters.xyrad = list(c(1, 1)),
  parameters.xyoff = list(c(1, 1)), parameters.minmax = c(0, 255),
  parameters.nbbin = 8, texture = "advanced", channel = NULL,
  verbose = FALSE)

\S4method{otbHaraTex}{RasterStack}(x, path_output = NULL,
  return_raster = TRUE, ram = "8192", parameters.xyrad = list(c(1, 1)),
  parameters.xyoff = list(c(1, 1)), parameters.minmax = c(0, 255),
  parameters.nbbin = 8, texture = "advanced", channel = NULL,
  verbose = FALSE)

\S4method{otbHaraTex}{character}(x, output_name = "hara",
  path_output = NULL, return_raster = FALSE, parameters.xyrad = list(c(1,
  1)), parameters.xyoff = list(c(1, 1)), parameters.minmax = c(0, 255),
  parameters.nbbin = 8, texture = "advanced", channel = NULL,
  verbose = FALSE, ram = "8192")
}
\arguments{
\item{x}{A \code{Raster*} object or a GeoTiff containing 1 or more gray 
value bands}

\item{return_raster}{boolean if TRUE a raster stack is returned}

\item{ram}{reserved memory in MB}

\item{parameters.xyrad}{list with the x and y radius in pixel indicating the kernel sizes for which the textures are calculated}

\item{parameters.xyoff}{vector containg the directional offsets. Valid combinations are: list(c(1,1),c(1,0),c(0,1),c(1,-1))}

\item{parameters.minmax}{minimum/maximum gray value which can occur.}

\item{parameters.nbbin}{number of gray level bins (classes)}

\item{texture}{type of filter "simple" "advanced" "higher"}

\item{channel}{sequence of bands to be processed}

\item{verbose}{switch for system messages default is FALSE}

\item{output_name}{string pattern vor individual naming of the output file(s)}

\item{n_grey}{Number of grey values.}

\item{parallel}{A logical value indicating whether parameters are calculated parallely or not}
}
\value{
A list of RasterStacks containing the texture parameters for each 
combination of channel and filter
}
\description{
Calculate selected texture parameters on gray level properties
}
\note{
the otb is used for filtering. please provide a GeoTiff file

More information at the texture tutorial site of
\link{http://www.fp.ucalgary.ca/mhallbey/more_informaton.htm}(Mryka Hall-Beyer)
Keep in mind that:\cr
Homogeneity is correlated with Contrast,  r = -0.80
Homogeneity is correlated with Dissimilarity, r = -0.95
GLCM Variance is correlated with Contrast,  r= 0.89
GLCM Variance is correlated with Dissimilarity,  r= 0.91
GLCM Variance is correlated with Homogeneity,  r= -0.83
Entropy is correlated with ASM,  r= -0.87
GLCM Mean and Correlation are more independent. For the same image,
GLCM Mean shows  r< 0.1 with any of the other texture measures demonstrated in this tutorial.
GLCM Correlation shows  r<0.5 with any other measure.
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbHaraTex(x=paste0(getwd(),"4490600_5321400.tif"),texture="simple")
}
}
\author{
Chris Reudenbach, Thomas Nauss
}

