% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otblocalStat.R
\name{otbLocalStat}
\alias{otbLocalStat}
\title{Calculates local statistics for a given kernel size}
\usage{
otbLocalStat(input = NULL, output_name = "localStat", ram = "8192",
  radius = 3, channel = NULL, return_raster = FALSE, path_output = NULL,
  verbose = FALSE)
}
\arguments{
\item{input}{of GeoTiff containing 1 ore more gray value bands}

\item{output_name}{string pattern for individual naming of the output file(s)}

\item{ram}{reserved memory in MB}

\item{radius}{computational window in pixel}

\item{channel}{sequence of bands to be processed}

\item{return_raster}{boolean if TRUE a raster stack is returned}

\item{verbose}{switch for system messages default is FALSE}
}
\description{
Calculate local statistics using OTB functions.
}
\note{
the otb is used for the calculation of the statistics. Please provide a GeoTiff file
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbLocalStat(input=paste0(getwd(),"4490600_5321400.tif"),radius=5)
}
}
\author{
Chris Reudenbach
}

