% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\docType{methods}
\name{pca}
\alias{pca}
\alias{pca,RasterStack-method}
\alias{pca,Satellite-method}
\title{Calculate image textures}
\usage{
\S4method{pca}{Satellite}(x, bcde, prfx = NULL, ...)

\S4method{pca}{RasterStack}(x, ...)
}
\arguments{
\item{x}{Satellite object, raster stack or raster layer.}

\item{bcde}{Band code(s) to be used for computing the image textures}

\item{...}{Further arguments passed on to \code{\link{glcm::glcm}}}

\item{mask}{raster layer with the same geometry as x with NA values at
positions where image textures should not be computed.}
}
\value{
If x is a Satellite object, a Satellite object with added image
textures; if x is a \code{raster::Raster*} object, a \code{raster::Raster*} 
object with converted layer(s).
}
\description{
Calculate image textures
}
\details{
The method is a wrapper for of glcm:glcm by A. Zvoleff
(see \url{https://cran.r-project.org/web/packages/glcm/index.html})
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.TIF"), full.names = TRUE)
sat <- satellite(files)

sat <- pca(sat, bcde = "B004n")
}
\references{
None.
}

