% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRasterFromPolygons.R
\name{sampleRasterFromPolygons}
\alias{sampleRasterFromPolygons}
\title{Sample raster values from polygons}
\arguments{
\item{x}{Satellite object, raster stack or raster layer.}
}
\value{
If x is a Satellite object, a Satellite object with added image
textures; if x is a \code{raster::Raster*} object, a \code{raster::Raster*} 
object with converted layer(s).
}
\description{
In the default settings, this is just a wrapper arround
raster::extract but with additional options in oder to select just a 
subsample of the extracted raster values based on properties of the size
of the polygons. The reason behind this is to be able to use e.g. 
training polygons for land-cover classifications of very different sizes 
without running into problems during model delevopment later.
}
\details{
The method is a wrapper for raster:extract with some extensions
}
\examples{
\dontrun{
}
}
\references{
None.
}

