% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satelliteTools-package.R
\docType{package}
\name{satelliteTools-package}
\alias{satelliteTools-package}
\alias{satelliteToolspackage}
\title{Smorgasboard for remote sensing functions.}
\description{
Smorgasboard for remote sensing functions
}
\details{
The package provides a variety of functions which are useful for 
computing higher level remote sensing products.
}
\references{
Some functions are taken and/or adopted from Sarah C. Goslee 
(2011). Analyzing Remote Sensing Data in R: The landsat Package. Journal of 
Statistical Software, 43(4), 1-25. Online available at 
\url{http://www.jstatsoft.org/v43/i04/}.
}
\author{
Thomas Nauss, Hanna Meyer, Florian Detsch, Tim Appelhans \cr
\cr
\emph{Maintainer:} Environmental Informatics \email{admin@environmentalinformatics-marburg.de}
}
\keyword{package}
