% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOTB.R
\name{searchOSgeo4WOTB}
\alias{searchOSgeo4WOTB}
\title{Search for valid OTB installations on a given windows drive}
\usage{
searchOSgeo4WOTB(dl = "C:")
}
\arguments{
\item{dl}{drive letter default is "C:"}
}
\value{
a dataframe with the OTB root dir the Version name and the installation type
}
\description{
provides a pretty good estimation of valid OTB installations on your Windows system
}
\examples{
\dontrun{
}
}
\author{
Chris Reudenbach
}
