% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOTB.R
\name{setOTBEnv}
\alias{setOTBEnv}
\title{Set environment parameters of OTB}
\usage{
setOTBEnv(path_otb = "C:\\\\OSGeo4W64\\\\bin",
  rootPathOtb = "C:\\\\OSGeo4W64")
}
\arguments{
\item{otbPath}{string contains path to otb binaries}

\item{sagaCmd}{string contains the full string to call otb launcher}
}
\value{
add otb pathes to the enviroment and creates global variables otbCmd
}
\description{
during a rsession you will have full access to OTB via the command line
}
\examples{
\dontrun{
}
}
