% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stIndices.R
\name{stIndices}
\alias{stIndices}
\title{Compute spectral remote sensing pixel-based indices}
\usage{
stIndices(sat, ...)
}
\arguments{
\item{sat}{a satellite::satellite class data set}
}
\value{
a raster::stack of the indices
}
\description{
Compute spectral remote sensing pixel-based indices
}
\examples{
\dontrun{
}
}

