% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stPCA.R
\name{stPCA}
\alias{stPCA}
\title{Compute principal component analysis}
\usage{
stPCA(sat, ...)
}
\arguments{
\item{sat}{a satellite::satellite class data set}
}
\value{
a raster::stack of the indices
}
\description{
Compute principal component analysis
}
\examples{
\dontrun{
}
}

