% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileRaster.R
\name{tileRaster}
\alias{tileRaster}
\title{Tile \code{Raster*} object with or without overlap}
\usage{
tileRaster(raster, size = -4, overlap = 0, outpath = NULL)
}
\arguments{
\item{raster}{\code{Raster*} object}

\item{size}{Size of the squared tiles (positive is pixel, negative number to be created)}

\item{overlap}{Overlap size (default NULL does not use overlap)}

\item{outpath}{Path of the output files (name will be equal to input name)}
}
\value{
None
}
\description{
Tile \code{Raster*} object using an optional overlap between the tiles.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
