% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeMatch.R
\name{timeMatch}
\alias{timeMatch}
\title{Match time stamps of two vectors}
\usage{
timeMatch(a, b)
}
\arguments{
\item{a}{Vector with time information}

\item{b}{Vector with time information}
}
\value{
A data frame with corresponding elements.
}
\description{
Match time stamps of two vectors in order to select the closest time in 
a second vector based on the time information in the first vector.
}
\examples{
\\notrun{
}

}
\author{
Thomas Nauss
}
\references{
NONE
}

