% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOTB.R
\name{initOTB}
\alias{initOTB}
\title{getOTB setup the orfeo toolbox bindings for an rsession}
\usage{
initOTB(defaultOTBPath = NULL,
  installationRoot = paste0(dirname(defaultOTBPath), "/"), otbType = NULL,
  DL = "C:")
}
\arguments{
\item{defaultOTBPath}{string contains path to otb binaries}

\item{DL}{hard drive letter}
}
\value{
add otb pathes to the enviroment and creates global variables otbPath
}
\description{
Try to find all valid OTB installation and returns the pathes and environment settings
}
\examples{
\dontrun{
## call it for a default OSGeo4W oinstallation of SAGA
initOTB("C:\\\\OSGeo4W64\\\\bin\\\\")
}

}

