% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otblocalStat.R
\docType{methods}
\name{otbLocalStat}
\alias{otbLocalStat}
\alias{otbLocalStat,RasterBrick-method}
\alias{otbLocalStat,RasterLayer-method}
\alias{otbLocalStat,RasterStack-method}
\alias{otbLocalStat,character-method}
\title{Calculates local statistics for a given kernel size}
\usage{
\S4method{otbLocalStat}{RasterBrick}(x, output_name = "locstat", radius = 3,
  path_output = NULL, return_raster = TRUE, channel = NULL,
  verbose = FALSE, ram = "8192")

\S4method{otbLocalStat}{RasterLayer}(x, output_name = "locstat", radius = 3,
  path_output = NULL, return_raster = TRUE, channel = NULL,
  verbose = FALSE, ram = "8192")

\S4method{otbLocalStat}{RasterStack}(x, output_name = "locstat", radius = 3,
  path_output = NULL, return_raster = TRUE, channel = NULL,
  verbose = FALSE, ram = "8192")

\S4method{otbLocalStat}{character}(x, output_name = "locstat", radius = 3,
  path_output = NULL, return_raster = FALSE, channel = NULL,
  verbose = FALSE, ram = "8192")
}
\arguments{
\item{x}{A \code{Raster*} object or a GeoTiff containing 1 or more gray 
value bands}

\item{output_name}{string pattern for individual naming of the output file(s)}

\item{radius}{computational window in pixel}

\item{return_raster}{boolean if TRUE a raster stack is returned}

\item{channel}{sequence of bands to be processed}

\item{verbose}{switch for system messages default is FALSE}

\item{ram}{reserved memory in MB}
}
\description{
Calculate local statistics using OTB functions.
}
\note{
the otb is used for the calculation of the statistics. Please provide a GeoTiff file
}
\examples{
\dontrun{
url<-"http://www.ldbv.bayern.de/file/zip/5619/DOP\%2040_CIR.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip(res,junkpaths = TRUE,overwrite = TRUE)
otbLocalStat(input=paste0(getwd(),"4490600_5321400.tif"),radius=5)
}
}
\author{
Chris Reudenbach, Thomas Nauss
}
