% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\docType{methods}
\name{pca}
\alias{pca}
\alias{pca,RasterStack-method}
\alias{pca,Satellite-method}
\alias{pca,list-method}
\title{Calculate PCA}
\usage{
\S4method{pca}{Satellite}(x, bcde, prfx = NULL, return_raster = FALSE, ...)

\S4method{pca}{RasterStack}(x, return_raster = FALSE, ...)

\S4method{pca}{list}(x, ignore_names = FALSE, center = TRUE, scale = TRUE,
  ...)
}
\arguments{
\item{x}{\code{Satellite}, \code{Raster*} or \code{list} of \code{Raster*} 
objects}

\item{bcde}{Band code(s) to be used for computing the pca}

\item{...}{Further arguments passed on to \code{\link{RStoolbox::rasterPCA}}
or \code{\link{stats::prcomp}}}

\item{ignore_names}{If list is supplied and if the band names of the 
individual rasters do not match (e.g. because of different observation dates),
ignore the names when building a single data frame for the PCA computation
(does not hurt).}

\item{center}{See \code{\link{stats::prcomp}}}

\item{scale}{See \code{\link{stats::prcomp}}}
}
\value{
If x is a Satellite object, a Satellite object with added image
textures; if x is a \code{raster::Raster*} object, a \code{raster::Raster*} 
object with converted layer(s).
}
\description{
Calculate principal components from raster data values. The
raster can either be supplied as \code{Raster*} object, \code{satellite*}
object or as a list of \code{Raster*} objects. The latter is usefull if the
raster dataset has been tiled into small observation areas using e.g. 
\code{\link{snipRaster}}.
}
\details{
The method for \code{Raster*} and \code{Satellite*} objects is basically a
wrapper arround the respective function in the RStoolbox package. For the list
(i.e. snip) related implementation, the stats::prcomp function is used to actually
compute the PCA.
}
\examples{
\dontrun{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.TIF"), full.names = TRUE)
sat <- satellite(files)

sat <- pca(sat, bcde = "B004n")
}
}
\references{
Benjamin Leutner and Ned Horning (2016). RStoolbox: Tools for Remote Sensing 
Data Analysis. R package version 0.1.6. 
https://CRAN.R-project.org/package=RStoolbox
}

