% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snipRaster.R
\name{snipRaster}
\alias{snipRaster}
\title{Tile \code{Raster*} object based on \code{Spatial*} object}
\usage{
snipRaster(raster, spatial, selector = NULL, buffer = NULL, byid = TRUE)
}
\arguments{
\item{raster}{\code{Raster*} object}

\item{spatial}{\code{Spatial*} object}

\item{selector}{Selector variable, if not NULL, only one raster tile per 
unique value of the selector variable is created (e.g. if one has multiple
observations per single geographical location).}

\item{buffer}{Buffer size (default NULL does not use buffer)}

\item{byid}{Parameter of \code{gBuffer} which is used in case a buffer should
be computed.}
}
\value{
A list of \code{Raster*} objects with one object per vector within variable
spatial.
}
\description{
Tile \code{Raster*} object based on \code{Spatial*} object using an optional 
buffer.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
