% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileRaster.R
\name{tileRaster}
\alias{tileRaster}
\title{Tile \code{Raster*} object with or without overlap}
\usage{
tileRaster(raster, tilenbr = c(2, 2), overlap = 0, outpath = NULL)
}
\arguments{
\item{raster}{\code{Raster*}. Object that should be tiled.}

\item{tilenbr}{'integer'. Vector given the number of tiles in c(x,y) direction.}

\item{overlap}{'integer'. Overlap in pixels between the individual tiles. Default NULL does not include any overlap.}

\item{outpath}{'character'. Path of the output files. Filenames will be taken from the \code{Raster*} layer names.}
}
\value{
None. Tiles are written to GeoTiff files in directory \code{outpath}.
}
\description{
Tile \code{Raster*} object using an optional overlap between the tiles.
}
\examples{
\\notrun{
tileRaster(raster = ndvi, tilenbr = c(5,4), overlap = 30, outpath = modis_tiles)
}

}
\references{
NONE
}
